refactor(backend): rename connection file and move to ilp
