+++ packages/backend/src/index.ts
@@ -37,7 +37,7 @@ import { createWebhookService } from './webhook/service'
 import { createConnectorService } from './payment-method/ilp/connector'
 import { createOpenAPI } from '@interledger/openapi'
 import { createAuthenticatedClient as createOpenPaymentsClient } from '@interledger/open-payments'
-import { createStreamCredentialsService } from './open_payments/connection/service'
+import { createStreamCredentialsService } from './payment-method/ilp/stream-credentials/service'
 import { createWalletAddressKeyService } from './open_payments/wallet_address/key/service'
 import { createReceiverService } from './open_payments/receiver/service'
 import { createRemoteIncomingPaymentService } from './open_payments/payment/incoming_remote/service'
