+++ packages/backend/src/open_payments/payment/incoming/model.test.ts
@@ -6,7 +6,7 @@ import { AppServices } from '../../../app'
 import { createIncomingPayment } from '../../../tests/incomingPayment'
 import { createWalletAddress } from '../../../tests/walletAddress'
 import { truncateTables } from '../../../tests/tableManager'
-import { IlpStreamCredentials } from '../../connection/service'
+import { IlpStreamCredentials } from '../../../payment-method/ilp/stream-credentials/service'
 import { serializeAmount } from '../../amount'
 import { IlpAddress } from 'ilp-packet'
 import { IncomingPayment } from './model'
